Organizing code into modules and updating variable names
